requirements.txt
django
djangorestframework
django-cors-headers
drf-yasg
psycopg2
Pillow
requests
djoser
mysqlclient
django-rest-swagger